+++ domain/src/main/java/daily/dayo/domain/usecase/post/RequestEditPostUseCase.kt
@@ -8,7 +8,7 @@ class RequestEditPostUseCase @Inject constructor(
     private val postRepository: PostRepository
 ) {
     suspend operator fun invoke(
-        postId: Int,
+        postId: Long,
         category: Category,
         contents: String,
         folderId: Int,
